+++ src/peripherals.rs
@@ -11,7 +11,7 @@ use crate::modem;
 #[cfg(any(esp32, esp32s2, esp32s3, esp32c6))]
 use crate::pcnt;
 use crate::rmt;
-#[cfg(all(esp_idf_soc_sdmmc_host_supported, feature = "experimental"))]
+#[cfg(esp_idf_soc_sdmmc_host_supported)]
 use crate::sd;
 use crate::spi;
 #[cfg(any(
@@ -86,9 +86,9 @@ pub struct Peripherals {
     #[cfg(any(all(esp32, esp_idf_eth_use_esp32_emac), esp_idf_eth_use_openeth))]
     pub mac: mac::MAC,
     pub modem: modem::Modem,
-    #[cfg(all(esp_idf_soc_sdmmc_host_supported, feature = "experimental"))]
+    #[cfg(esp_idf_soc_sdmmc_host_supported)]
     pub sdmmc0: sd::mmc::SDMMC0,
-    #[cfg(all(esp_idf_soc_sdmmc_host_supported, feature = "experimental"))]
+    #[cfg(esp_idf_soc_sdmmc_host_supported)]
     pub sdmmc1: sd::mmc::SDMMC1,
     #[cfg(all(esp_idf_soc_temp_sensor_supported, esp_idf_version_major = "5"))]
     pub temp_sensor: temp_sensor::TempSensor,
@@ -191,9 +191,9 @@ impl Peripherals {
             #[cfg(any(all(esp32, esp_idf_eth_use_esp32_emac), esp_idf_eth_use_openeth))]
             mac: mac::MAC::new(),
             modem: modem::Modem::new(),
-            #[cfg(all(esp_idf_soc_sdmmc_host_supported, feature = "experimental"))]
+            #[cfg(esp_idf_soc_sdmmc_host_supported)]
             sdmmc0: sd::mmc::SDMMC0::new(),
-            #[cfg(all(esp_idf_soc_sdmmc_host_supported, feature = "experimental"))]
+            #[cfg(esp_idf_soc_sdmmc_host_supported)]
             sdmmc1: sd::mmc::SDMMC1::new(),
             #[cfg(all(esp_idf_soc_temp_sensor_supported, esp_idf_version_major = "5"))]
             temp_sensor: temp_sensor::TempSensor::new(),